Refactor | Dead Custom Provider Code
