+++ .env.example
@@ -16,6 +16,3 @@ POSTGRES_PORT=5432
 POSTGRES_USER=your_username
 POSTGRES_PASSWORD=your_password
 POSTGRES_DB=tomodb
-
-# Optional: in-place persona workflow button timeout in milliseconds (default: 120000).
-PERSONA_WORKFLOW_COMPONENT_TIMEOUT_MS=120000
