+++ src/commands/provider/remove.ts
@@ -12,7 +12,8 @@ import { log, ColorCode } from "@/utils/misc/logger";
 import { replyInfoEmbed } from "@/utils/discord/ui/embeds";
 import type { UserRow, ErrorContext } from "@/types/db/schema";
 import { getProviderDisplayName } from "@/utils/provider/providerInfoRegistry";
-import { isCustomProvider, deleteCustomLLMEntry } from "@/utils/discord/customProviderModal";
+import { deleteCustomLLMEntry } from "@/utils/provider/legacyCustomProvider";
+import { isCustomProvider } from "@/utils/provider/customProviderUtils";
 import { hasRegisteredCustomProvider, loadProviderDefaultSelectionIds } from "@/utils/provider/savedProviderConfig";
 import { cleanupCustomProviderArtifacts } from "@/utils/provider/customEndpointService";
 import { promptForSavedProvider } from "@/utils/discord/providerPicker";
