+++ src/locales/en-US/commands/tool.ts
@@ -148,9 +148,9 @@ I have built-in features to help reduce costs from abusers or spammers in your s
       add_to_docs_error_description: `Could not save the summary to document memory.`,
     },
     refresh: {
-      description: `Clears the recent conversation history.`,
+      description: `Clear conversation history (this channel only).`,
       title: `🧹 Conversation History Cleared`,
-      response: `All messages above this one will now be ignored.\n This channel's STM has also been cleared.`,
+      response: `All messages above this one will now be ignored.\n This channel's [STM](https://docs.tomoribot.app/features/knowledge/memory/#short-term-memory-stm) has also been cleared.`,
       footer: `Delete this embed to allow older messages above to be included again.`,
     },
     status: {
