+++ src/locales/ja/commands/tool.ts
@@ -150,7 +150,7 @@ export default {
     refresh: {
       description: `最近の会話履歴をクリアします。`,
       title: `🧹 会話履歴がクリアされました`,
-      response: `コンテキストがリフレッシュされました。これより上のすべてのメッセージは無視されます。`,
+      response: `これより上のすべてのメッセージは無視されます。\n このチャンネルの[STM](https://docs.tomoribot.app/features/knowledge/memory/#short-term-memory-stm)もクリアされました。`,
       footer: `この埋め込みを削除すると、以前のメッセージがもう一度含まれるようになります。`,
     },
     status: {
